py_server/ocr_svr/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
easyocr==1.7.0
torch==2.1.0+cpu
torchvision==0.16.0+cpu
pillow==9.5.0
numpy==1.24.3
python-multipart==0.0.6
opencv-python==4.8.0.76 